Moved label class inside classes directory
