--- lib/label.rb
@@ -1,18 +1,0 @@
-require_relative '../lib/item'
-
-class Label
-  attr_accessor :title, :color
-  attr_reader :id, :items
-
-  def initialize(title, color)
-    @id = Random.rand(1..10_000)
-    @title = title
-    @color = color
-    @items = []
-  end
-
-  def add_item(item)
-    @items << item unless @items.include?(item)
-    item.add_label = self
-  end
-end
